[Autofill] Extract FormFieldData::max_length of textareas
Not only text-control <input> elements but also <textarea> supports
the maxlength attribute. This CL sets FormFieldData::max_length also
for textareas.

This CL changes the default maxiumm length for text-control <input>
and <textarea> elements to the maximum 32 bit number. Before this
CL, it was the maximum int, whose size isn't specified.

The reason that this CL chooses the maximum 32 bit number and not
64 bit, is to avoid the risk of integer overflows in the, for
example, the rationalization code.

Therefore, this CL moves the maxlength default value from
WebInputElement to form_autofill_util.cc.

Bug: 1484810
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4874062
Commit-Queue: Christoph Schwering <schwering@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Auto-Submit: Christoph Schwering <schwering@google.com>
Cr-Commit-Position: refs/heads/main@{#1205150}

+++ components/password_manager/ios/shared_password_controller_unittest.mm
@@ -152,7 +152,7 @@ class MockPasswordGenerationFrameHelper : public PasswordGenerationFrameHelper {
               (const GURL&,
                autofill::FormSignature,
                autofill::FieldSignature,
-               uint32_t),
+               uint64_t),
               (override));
 
   MOCK_METHOD(bool, IsGenerationEnabled, (bool), (const));
